+++ tests/ovp/ngx/projects/kaltura-ngx-client/src/tests/kaltura-file-request.spec.ts
@@ -1,8 +1,9 @@
-import {ThumbAssetServeAction} from "../api/types/ThumbAssetServeAction";
-import {getClient} from "./utils";
-import {LoggerSettings, LogLevels} from "../api/kaltura-logger";
-import {KalturaClient} from "../kaltura-client.service";
-import "rxjs/add/observable/throw";
+import {ThumbAssetServeAction} from "../lib/api/types/ThumbAssetServeAction";
+import { asyncAssert, getClient } from "./utils";
+import {LoggerSettings, LogLevels} from "../lib/api/kaltura-logger";
+import {KalturaClient} from "../lib/kaltura-client.service";
+import { environment } from '../lib/environment';
+
 
 describe("Kaltura File request", () => {
   let kalturaClient: KalturaClient = null;
@@ -29,14 +30,19 @@ describe("Kaltura File request", () => {
       thumbAssetId: "1_ep9epsxy"
     });
 
-    kalturaClient.setDefaultRequestOptions({ks: "YWIyZDAxYWRhZmQ1NzhjMzQ5ZmI3Nzc4MzVhYTJkMGI1NDdhYzA5YnwxNzYzMzIxOzE3NjMzMjE7MTUxMjA1MzA1MzsyOzE1MTE5NjY2NTMuNTk7YWRtaW47ZGlzYWJsZWVudGl0bGVtZW50Ozs"});
+    kalturaClient.setDefaultRequestOptions({ks: "ks123"});
+
+    expect.assertions(3);
 
     kalturaClient.request(thumbRequest)
       .subscribe(
         result => {
-          expect(result).toBeDefined();
-          expect(result.url).toBeDefined();
-          expect(result.url).toBe("https://www.kaltura.com/api_v3/service/thumbasset/action/serve?clientTag=ngx-tests?apiVersion=3.3.0&ks=YWIyZDAxYWRhZmQ1NzhjMzQ5ZmI3Nzc4MzVhYTJkMGI1NDdhYzA5YnwxNzYzMzIxOzE3NjMzMjE7MTUxMjA1MzA1MzsyOzE1MTE5NjY2NTMuNTk7YWRtaW47ZGlzYWJsZWVudGl0bGVtZW50Ozs&thumbAssetId=1_ep9epsxy&format=1");
+          asyncAssert(() => {
+            expect(result).toBeDefined();
+            expect(result.url).toBeDefined();
+            expect(result.url).toBe(`https://www.kaltura.com/api_v3/service/thumbasset/action/serve?clientTag=ngx-tests&ks=ks123&thumbAssetId=1_ep9epsxy&apiVersion=${environment.request.apiVersion}&format=1`);
+
+          });
 
           done();
         },
@@ -52,15 +58,19 @@ describe("Kaltura File request", () => {
       thumbAssetId: "thumbAssetId"
     });
 
+    expect.assertions(3);
+
     kalturaClient.multiRequest([thumbRequest])
       .subscribe(
         result => {
-          fail("got response instead of error");
+          done.fail("got response instead of error");
         },
         error => {
-          expect(error).toBeDefined();
-          expect(error).toBeInstanceOf(Error);
-          expect(error.message).toBe("multi-request not support requests of type 'KalturaFileRequest', use regular request instead");
+          asyncAssert(() => {
+            expect(error).toBeDefined();
+            expect(error).toBeInstanceOf(Error);
+            expect(error.message).toBe("multi-request not support requests of type 'KalturaFileRequest', use regular request instead");
+          });
           done();
         });