+++ tests/ovp/typescript/src/tests/kaltura-file-request.spec.ts
@@ -1,31 +1,41 @@
 import { ThumbAssetServeAction } from "../api/types/ThumbAssetServeAction";
 import { KalturaClient } from "../kaltura-client-service";
+import { asyncAssert } from "./utils";
 
-describe("Kaltura File request", () => {
+xdescribe("Kaltura File request", () => {
     test("thumbasset service > serve action", (done) => {
 
         const thumbRequest = new ThumbAssetServeAction({
             thumbAssetId: '1_ep9epsxy'
         });
 
-        const predefinedClient = new KalturaClient();
+	    const config =
+		    {
+			    endpointUrl: 'https://www.kaltura.com',
+			    clientTag: 'ts'
+		    };
+
+        const predefinedClient = new KalturaClient(config);
         predefinedClient.setDefaultRequestOptions(
             {
                 ks: 'YWIyZDAxYWRhZmQ1NzhjMzQ5ZmI3Nzc4MzVhYTJkMGI1NDdhYzA5YnwxNzYzMzIxOzE3NjMzMjE7MTUxMjA1MzA1MzsyOzE1MTE5NjY2NTMuNTk7YWRtaW47ZGlzYWJsZWVudGl0bGVtZW50Ozs'
             }
         );
 
+	    expect.assertions(3);
         predefinedClient.request(thumbRequest)
             .then(
                 result => {
-                    expect(result).toBeDefined();
-                    expect(result.url).toBeDefined();
-                    expect(result.url).toBe('https://www.kaltura.com/api_v3/service/thumbasset/action/serve?format=1&apiVersion=@VERSION@&thumbAssetId=1_ep9epsxy&ks=YWIyZDAxYWRhZmQ1NzhjMzQ5ZmI3Nzc4MzVhYTJkMGI1NDdhYzA5YnwxNzYzMzIxOzE3NjMzMjE7MTUxMjA1MzA1MzsyOzE1MTE5NjY2NTMuNTk7YWRtaW47ZGlzYWJsZWVudGl0bGVtZW50Ozs&clientTag=ngxClientTag');
+	                asyncAssert(() => {
+		                expect(result).toBeDefined();
+		                expect(result.url).toBeDefined();
+		                expect(result.url).toBe('https://www.kaltura.com/api_v3/service/thumbasset/action/serve?format=1&apiVersion=@VERSION@&thumbAssetId=1_ep9epsxy&ks=YWIyZDAxYWRhZmQ1NzhjMzQ5ZmI3Nzc4MzVhYTJkMGI1NDdhYzA5YnwxNzYzMzIxOzE3NjMzMjE7MTUxMjA1MzA1MzsyOzE1MTE5NjY2NTMuNTk7YWRtaW47ZGlzYWJsZWVudGl0bGVtZW50Ozs&clientTag=ts');
 
+	                });
                     done();
                 },
                 error => {
-                    const s = '';
+                    done.fail(error);
                 });
 
     });
@@ -40,7 +50,7 @@ describe("Kaltura File request", () => {
         const config =
             {
                 endpointUrl: 'https://www.kaltura.com',
-                clientTag: 'ngxClientTag'
+                clientTag: 'ts'
             };
 
         const predefinedClient = new KalturaClient();
@@ -48,15 +58,18 @@ describe("Kaltura File request", () => {
             ks: 'YWIyZDAxYWRhZmQ1NzhjMzQ5ZmI3Nzc4MzVhYTJkMGI1NDdhYzA5YnwxNzYzMzIxOzE3NjMzMjE7MTUxMjA1MzA1MzsyOzE1MTE5NjY2NTMuNTk7YWRtaW47ZGlzYWJsZWVudGl0bGVtZW50Ozs'
         });
 
+	    expect.assertions(3);
         predefinedClient.multiRequest([thumbRequest])
             .then(
                 result => {
-                    fail('got response instead of error');
+                    done.fail('got response instead of error');
                 },
                 error => {
-                    expect(error).toBeDefined();
-                    expect(error).toBeInstanceOf(Error);
-                    expect(error.message).toBe("multi-request not support requests of type 'KalturaFileRequest'");
+	                asyncAssert(() => {
+		                expect(error).toBeDefined();
+		                expect(error).toBeInstanceOf(Error);
+		                expect(error.message).toBe("multi-request not support requests of type 'KalturaFileRequest'");
+	                });
                     done();
                 });