+++ sources/typescript/CHANGELOG.md
@@ -2,6 +2,35 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 6.4.2 (2018-11-29)
+
+### Fix
+
+* typescript declaration issue for error response
+* async unit-tests 
+
+
+## 6.4.1 (2018-11-28)
+
+### Fix
+
+* cancel request now the cancel underline xhr request
+
+### Fix
+
+* use user custom chunk size if provided  
+
+
+## 6.4.0 (2018-11-18)
+
+### Features
+
+* update minimal custom chunk size to 100Kb
+
+### Fix
+
+* use user custom chunk size if provided  
+
 ## 6.3.2 (2018-09-13)
 
 ### Fix